requirements.txt
flask~=3.1.0
pymongo~=4.8.0
pandas~=2.2.1
unidecode~=1.3.8
python-dotenv~=1.0.1
jsonpickle~=3.2.2
PyPDF2~=3.0.1
dash~=2.18.1
plotly~=5.13.1
werkzeug~=3.1.3
openai~=1.35.10
cheroot~=10.0.1
langchain-community~=0.2.6
langchain-openai~=0.1.14
sphinx-rtd-theme~=3.0.2